#199 feat: 리뷰 삭제 수정
리뷰 조회 및 리뷰 삭제 기능 구현

+++ src/main/java/kr/oshino/eataku/review/user/repository/ReviewRepository.java
@@ -9,8 +9,9 @@ import java.util.List;
 
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-    void deleteByReviewNo(int reviewNo);
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
+
+    void deleteByReviewNo(Integer reviewNo);
 
     int countByMember(Member toMember);
